refactor(ABHI-1360): split gh token CLI and dedupe security tests
Move CLI helpers into src/utils/gh_token_cli.py, keep scripts/gh_token_env.py
as a thin entrypoint, and consolidate repeated tempfile setup in tests to
address CodeScene code health findings.
