+++ tests/test_gh_token_env.py
@@ -6,7 +6,9 @@ import subprocess
 import sys
 import tempfile
 import unittest
+from contextlib import contextmanager
 from pathlib import Path
+from typing import Iterator
 
 ROOT = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT))
@@ -20,71 +22,48 @@ SECURE_AUTOMATION_SCRIPTS = (
 from src.utils.env_file_parser import EnvParseError, parse_env_file
 
 
+@contextmanager
+def temporary_env_file(content: str) -> Iterator[Path]:
+    with tempfile.NamedTemporaryFile("w", delete=False) as handle:
+        handle.write(content)
+        path = Path(handle.name)
+    try:
+        yield path
+    finally:
+        path.unlink()
+
+
 class TestGhTokenEnvParser(unittest.TestCase):
     def test_parses_simple_assignment(self) -> None:
-        with tempfile.NamedTemporaryFile("w", delete=False) as handle:
-            handle.write("GH_TOKEN=abc123\n")
-            path = Path(handle.name)
-
-        try:
+        with temporary_env_file("GH_TOKEN=abc123\n") as path:
             self.assertEqual(parse_env_file(path), {"GH_TOKEN": "abc123"})
-        finally:
-            path.unlink()
 
     def test_rejects_command_injection_line(self) -> None:
-        with tempfile.NamedTemporaryFile("w", delete=False) as handle:
-            handle.write('GH_TOKEN=safe\n$(touch /tmp/pwned)\n')
-            path = Path(handle.name)
-
-        try:
+        with temporary_env_file('GH_TOKEN=safe\n$(touch /tmp/pwned)\n') as path:
             with self.assertRaises(EnvParseError):
                 parse_env_file(path)
-        finally:
-            path.unlink()
 
     def test_rejects_shell_command_disguised_as_key(self) -> None:
-        with tempfile.NamedTemporaryFile("w", delete=False) as handle:
-            handle.write("GH_TOKEN=$(id)\n")
-            path = Path(handle.name)
-
-        try:
+        with temporary_env_file("GH_TOKEN=$(id)\n") as path:
             with self.assertRaises(EnvParseError):
                 parse_env_file(path)
-        finally:
-            path.unlink()
 
     def test_allows_quoted_token_values(self) -> None:
-        with tempfile.NamedTemporaryFile("w", delete=False) as handle:
-            handle.write('GH_TOKEN="ghp_abc-def_123"\n')
-            path = Path(handle.name)
-
-        try:
+        with temporary_env_file('GH_TOKEN="ghp_abc-def_123"\n') as path:
             self.assertEqual(
                 parse_env_file(path), {"GH_TOKEN": "ghp_abc-def_123"}
             )
-        finally:
-            path.unlink()
 
     def test_strips_trailing_inline_comments_from_unquoted_values(self) -> None:
-        with tempfile.NamedTemporaryFile("w", delete=False) as handle:
-            handle.write("GH_TOKEN=abc123  # local token\n")
-            path = Path(handle.name)
-
-        try:
+        with temporary_env_file("GH_TOKEN=abc123  # local token\n") as path:
             self.assertEqual(parse_env_file(path), {"GH_TOKEN": "abc123"})
-        finally:
-            path.unlink()
 
     def test_cli_get_does_not_execute_malicious_file(self) -> None:
         marker = Path(tempfile.gettempdir()) / "gh_token_env_pwned_marker"
         if marker.exists():
             marker.unlink()
 
-        with tempfile.NamedTemporaryFile("w", delete=False) as handle:
-            handle.write(f'GH_TOKEN=ok\n$(touch "{marker}")\n')
-            path = Path(handle.name)
-
-        try:
+        with temporary_env_file(f'GH_TOKEN=ok\n$(touch "{marker}")\n') as path:
             proc = subprocess.run(
                 [sys.executable, str(HELPER), "--get", "GH_TOKEN", str(path)],
                 capture_output=True,
@@ -93,10 +72,6 @@ class TestGhTokenEnvParser(unittest.TestCase):
             )
             self.assertNotEqual(proc.returncode, 0)
             self.assertFalse(marker.exists())
-        finally:
-            path.unlink()
-            if marker.exists():
-                marker.unlink()
 
 
 class TestAutomationScripts(unittest.TestCase):
